requirements.txt
gradio
langgraph
langchain
langchain-openai
langchain-community
langchain-experimental
langchain-text-splitters
langchain-core
langchain-chroma
playwright
python-dotenv
pydantic
requests
typing-extensions
chromadb
nest-asyncio
beautifulsoup4
lxml
PyPDF2
pdfplumber
pypdf

